+++ test/Braintree.TestUtil/TestHelper.cs
@@ -51,7 +51,7 @@ namespace Braintree.TestUtil
                     ""postal_code"": ""94112""
                 },
                 ""account_type"": ""checking"",
-                ""routing_number"": ""123456789"",
+                ""routing_number"": ""021000021"",
                 ""account_number"": ""567891234"",
                 ""account_holder_name"": ""Dan Schulman"",
                 ""account_description"": ""PayPal Checking - 1234"",